Revert grammar to what was shipped in VS Code
This change reverts the TextMate grammar to the state it was shipped
as in Visual Studio Code before they started sourcing this repository
for the grammar file.  We're making this change to avoid introducing
regressions into the grammar that ships with VS Code until V2 of the
PowerShell grammar is complete.